+++ coordinator/helper.go
@@ -14,6 +14,9 @@
 package coordinator
 
 import (
+	"strings"
+
+	"github.com/pingcap/ticdc/pkg/config"
 	"github.com/pingcap/ticdc/pkg/messaging"
 )
 
@@ -35,7 +38,54 @@ const (
 	EventPeriod
 )
 
+const changefeedErrorMetricMsgLimit = 256
+
 type Event struct {
 	eventType int
 	message   *messaging.TargetMessage
+}
+
+type changefeedErrorMetricLabels struct {
+	keyspace   string
+	changefeed string
+	state      string
+	code       string
+	message    string
+}
+
+func (l changefeedErrorMetricLabels) labelValues() []string {
+	return []string{l.keyspace, l.changefeed, l.state, l.code, l.message}
+}
+
+func normalizeChangefeedErrorMetricMessage(message string) string {
+	message = strings.Join(strings.Fields(message), " ")
+	if len(message) <= changefeedErrorMetricMsgLimit {
+		return message
+	}
+	return message[:changefeedErrorMetricMsgLimit-3] + "..."
+}
+
+func getChangefeedErrorMetricLabels(info *config.ChangeFeedInfo) (changefeedErrorMetricLabels, bool) {
+	if info == nil {
+		return changefeedErrorMetricLabels{}, false
+	}
+	if info.State != config.StateFailed && info.State != config.StateWarning {
+		return changefeedErrorMetricLabels{}, false
+	}
+
+	runningErr := info.Error
+	if runningErr == nil {
+		runningErr = info.Warning
+	}
+	if runningErr == nil {
+		return changefeedErrorMetricLabels{}, false
+	}
+
+	return changefeedErrorMetricLabels{
+		keyspace:   info.ChangefeedID.Keyspace(),
+		changefeed: info.ChangefeedID.Name(),
+		state:      string(info.State),
+		code:       runningErr.Code,
+		message:    normalizeChangefeedErrorMetricMessage(runningErr.Message),
+	}, true
 }
